gasffsdfas

+++ app/Http/Controllers/ProductController.php
@@ -2,10 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\UserProducts;
+use Exception;
+use App\Models\User;
 use Inertia\Inertia;
 use App\Models\Product;
-use App\Models\User;
+use App\Models\UserProducts;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Route;
@@ -25,23 +26,23 @@ class ProductController extends Controller
 
             if ($user->hasRole('admin')) {
                 $products = UserProducts::select('products.product_id', 'products.product_name', 'products.product_img', 'products.product_price', 'products.product_img', 'products.product_quantity', 'users.name', 'users.id')
-                ->join('products', 'user_products.product_id', '=', 'products.product_id')
-                ->join('users', 'user_products.user_id', '=', 'users.id')
-                ->orderByDesc('product_id')
-                ->when($request->searchTerm, function ($query, $searchTerm) {
-                    return $query->where('product_name', 'like', '%' . $searchTerm . '%');
-                })->paginate(5);
+                    ->join('products', 'user_products.product_id', '=', 'products.product_id')
+                    ->join('users', 'user_products.user_id', '=', 'users.id')
+                    ->orderByDesc('product_id')
+                    ->when($request->searchTerm, function ($query, $searchTerm) {
+                        return $query->where('product_name', 'like', '%' . $searchTerm . '%');
+                    })->paginate(5);
             } elseif ($user->hasRole('editor')) {
                 $products = UserProducts::select('products.product_id', 'products.product_name', 'products.product_img', 'products.product_price', 'products.product_img', 'products.product_quantity', 'users.name', 'users.id')
-                ->join('products', 'user_products.product_id', '=', 'products.product_id')
-                ->join('users', 'user_products.user_id', '=', 'users.id')
-                ->orderByDesc('product_id')
-                ->when($request->searchTerm, function ($query, $searchTerm) {
-                    return $query->where('product_name', 'like', '%' . $searchTerm . '%');
-                })->where('users.id', $user_id)
-                ->paginate(5);
+                    ->join('products', 'user_products.product_id', '=', 'products.product_id')
+                    ->join('users', 'user_products.user_id', '=', 'users.id')
+                    ->orderByDesc('product_id')
+                    ->when($request->searchTerm, function ($query, $searchTerm) {
+                        return $query->where('product_name', 'like', '%' . $searchTerm . '%');
+                    })->where('users.id', $user_id)
+                    ->paginate(5);
             }
-            
+
 
             return Inertia::render('Products/Products', [
                 'products' => $products,
@@ -85,7 +86,7 @@ class ProductController extends Controller
                 return Redirect::route('products.index')
                     ->with('successMessage', 'Product Added Successfully');
             }
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             return Redirect::back()->with('errorMessage', 'Error adding product: ' . $e->getMessage())->withInput();
         }
     }
@@ -93,13 +94,33 @@ class ProductController extends Controller
     public function destroy($product_id)
     {
         try {
+            $user = auth()->user();
             $product = Product::find($product_id);
+            $user_product = UserProducts::where('product_id', $product_id)->first();
             $product_name = $product->product_name;
-            $product->delete();
 
-            return Redirect::route('products.index')
-                ->with('successMessage', 'Product: ' . $product_name . ' Deleted Successfully');
-        } catch (\Exception $e) {
+            if ($user->hasRole('admin')) {
+                $product->delete();
+
+                if ($product->product_img !== null && storage::exists('public/' . $product->product_img)) {
+                    Storage::delete('public/' . $product->product_img);
+                }
+                return Redirect::route('products.index')
+                    ->with('successMessage', 'Product: ' . $product_name . ' Deleted Successfully');
+            } elseif ($user->hasRole('editor')) {
+                if ($user->id === $user_product->user_id) {
+                    $product->delete();
+
+                    if ($product->product_img !== null && storage::exists('public/' . $product->product_img)) {
+                        Storage::delete('public/' . $product->product_img);
+                    }
+                    return Redirect::route('products.index')
+                        ->with('successMessage', 'Product: ' . $product_name . ' Deleted Successfully');
+                } else {
+                    throw new Exception("You don't have access to this product");
+                }
+            }
+        } catch (Exception $e) {
             return Redirect::back()->with('errorMessage', 'Error deleting product: ' . $e->getMessage())->withInput();
         }
 
@@ -108,12 +129,24 @@ class ProductController extends Controller
     public function getProduct($product_id)
     {
         try {
+            $user = auth()->user();
             $product = Product::find($product_id);
+            $user_product = UserProducts::where('product_id', $product_id)->first();
 
-            return Inertia::render('Products/ViewProduct', [
-                'product' => $product,
-            ]);
-        } catch (\Exception $e) {
+            if ($user->hasRole('admin')) {
+                return Inertia::render('Products/ViewProduct', [
+                    'product' => $product,
+                ]);
+            } elseif ($user->hasRole('editor')) {
+                if ($user->id === $user_product->user_id) {
+                    return Inertia::render('Products/ViewProduct', [
+                        'product' => $product,
+                    ]);
+                } else {
+                    throw new Exception("You don't have access to this product");
+                }
+            }
+        } catch (Exception $e) {
             return Redirect::back()->with('errorMessage', 'Error getting product: ' . $e->getMessage())->withInput();
         }
     }
@@ -121,7 +154,9 @@ class ProductController extends Controller
     public function updateProduct(Request $request, $product_id)
     {
         try {
+            $user = auth()->user();
             $product = Product::find($product_id);
+            $user_product = UserProducts::where('product_id', $product_id)->first();
             $product_name = $product->product_name;
 
             $validator = Validator::make($request->all(), [
@@ -130,25 +165,42 @@ class ProductController extends Controller
                 'product_quantity' => 'required|integer|min:1',
             ]);
 
+            if ($user->hasRole('admin')) {
+                if ($validator->fails()) {
+                    return Redirect::route('products.show', ['product_id' => $product_id])
+                        ->withErrors($validator);
+                } else {
+                    $product->product_name = $request->input('product_name');
+                    $product->product_price = $request->input('product_price');
+                    $product->product_quantity = $request->input('product_quantity');
+                    $product->updated_at = now();
 
-            if (!$product) {
-                return Redirect::route('products.show', ['product_id' => $product_id])
-                    ->with('errorMessage', 'Product no found');
-            } elseif ($validator->fails()) {
-                return Redirect::route('products.show', ['product_id' => $product_id])
-                    ->withErrors($validator);
-            } else {
-                $product->product_name = $request->input('product_name');
-                $product->product_price = $request->input('product_price');
-                $product->product_quantity = $request->input('product_quantity');
-                $product->updated_at = now();
+                    $product->save();
 
-                $product->save();
+                    return Redirect::route('products.show', ['product_id' => $product_id])
+                        ->with('successMessage', 'Product: ' . $product_name . ' Updated Successfully');
+                }
+            } elseif ($user->hasRole('editor')) {
+                if ($user->id === $user_product->user_id) {
+                    if ($validator->fails()) {
+                        return Redirect::route('products.show', ['product_id' => $product_id])
+                            ->withErrors($validator);
+                    } else {
+                        $product->product_name = $request->input('product_name');
+                        $product->product_price = $request->input('product_price');
+                        $product->product_quantity = $request->input('product_quantity');
+                        $product->updated_at = now();
 
-                return Redirect::route('products.show', ['product_id' => $product_id])
-                    ->with('successMessage', 'Product: ' . $product_name . ' Updated Successfully');
+                        $product->save();
+
+                        return Redirect::route('products.show', ['product_id' => $product_id])
+                            ->with('successMessage', 'Product: ' . $product_name . ' Updated Successfully');
+                    }
+                } else {
+                    throw new Exception("You don't have access to this product");
+                }
             }
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             return Redirect::back()->with('errorMessage', 'Error updating product: ' . $e->getMessage())->withInput();
         }
     }
